move` Logo` into `atoms`
