+++ web/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Logo } from "./components/Logo";
+import { Logo } from "./components/atoms/Logo";
 
 export const App = () => {
   return (
