+++ web/src/components/atoms/Logo.tsx
